Avoid dropping duplicates when calculating scores awarded wrt gender
Since the table doesn't include the 'from_code2' column it's important
not to drop duplicates. Dropping duplicates would mean merging all rows
where the same country was given the same score in the same year by
multiple countries. However, this is not an unusual circumstance, and
the votes should be counted separately in this case (rather than being
merged into a single vote).